refactor: HandlerMappings 일급 컬렉션 생성
