+++ app/src/main/java/com/techcourse/DispatcherServletInitializer.java
@@ -7,6 +7,8 @@ import web.org.springframework.web.WebApplicationInitializer;
 import webmvc.org.springframework.web.servlet.mvc.tobe.adapter.ControllerHandlerAdapter;
 import webmvc.org.springframework.web.servlet.mvc.tobe.adapter.HandlerAdapters;
 import webmvc.org.springframework.web.servlet.mvc.tobe.adapter.HandlerExecutionHandlerAdapter;
+import webmvc.org.springframework.web.servlet.mvc.tobe.mapping.AnnotationHandlerMapping;
+import webmvc.org.springframework.web.servlet.mvc.tobe.mapping.HandlerMappings;
 
 /**
  * Base class for {@link WebApplicationInitializer}
@@ -17,12 +19,14 @@ public class DispatcherServletInitializer implements WebApplicationInitializer {
     private static final Logger log = LoggerFactory.getLogger(DispatcherServletInitializer.class);
     private static final String DEFAULT_SERVLET_NAME = "dispatcher";
 
+    private final HandlerMappings handlerMappings = new HandlerMappings();
     private final HandlerAdapters handlerAdapters = new HandlerAdapters();
 
     @Override
     public void onStartup(final ServletContext servletContext) {
+        initHandlerMappings();
         initHandlerAdapters();
-        final var dispatcherServlet = new DispatcherServlet(handlerAdapters);
+        final var dispatcherServlet = new DispatcherServlet(handlerMappings, handlerAdapters);
 
         final var registration = servletContext.addServlet(DEFAULT_SERVLET_NAME, dispatcherServlet);
         if (registration == null) {
@@ -34,6 +38,11 @@ public class DispatcherServletInitializer implements WebApplicationInitializer {
         registration.addMapping("/");
 
         log.info("Start AppWebApplication Initializer");
+    }
+
+    private void initHandlerMappings() {
+        handlerMappings.addHandlerMapping(new ManualHandlerMapping());
+        handlerMappings.addHandlerMapping(new AnnotationHandlerMapping("com.techcourse.controller"));
     }
 
     private void initHandlerAdapters() {
